+++ public/sw.js
@@ -1,1 +1,1 @@
-if(!self.define){let e,a={};const s=(s,i)=>(s=new URL(s+".js",i).href,a[s]||new Promise((a=>{if("document"in self){const e=document.createElement("script");e.src=s,e.onload=a,document.head.appendChild(e)}else e=s,importScripts(s),a()})).then((()=>{let e=a[s];if(!e)throw new Error(`Module ${s} didn’t register its module`);return e})));self.define=(i,n)=>{const c=e||("document"in self?document.currentScript.src:"")||location.href;if(a[c])return;let r={};const t=e=>s(e,c),o={module:{uri:c},exports:r,require:t};a[c]=Promise.all(i.map((e=>o[e]||t(e)))).then((e=>(n(...e),r)))}}define(["./workbox-c06b064f"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/static/chunks/231-8cbc9f07fd7974e2.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/2433-a3c2272ce6e32cc3.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/2778-6a1712d23e611409.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/2967-86aef47a05f0281b.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/3771-c2dc1aac67d0345a.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/5190-88107600b5c8b2fc.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/6079-79ce56e79ed4ddb4.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/6111-fc803a4b575d4b6c.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/6612-1dc1b3cadee020c9.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/7116-60045fc3c0b0381f.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/7271-938511c36ff30d26.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/8173-275b4573b7cac1f5.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/8318a8be-eb73eb9ebc18b091.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/8866-0a1ff8780df5aa87.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/9646-15f0506087ab33c8.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/998-3aa0451c39cf29b9.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/(common)/(auth)/signin/page-211b7bed5ceedb6e.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/(common)/(auth)/signup/page-b3a0a794cf008435.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/(common)/explore/page-efaab558ea3a5977.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/(common)/ranking/page-a03321b36ce5c331.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/(common)/selectrole/page-b78e3f88cc670408.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/_not-found/page-c51cc89ad3b69e1f.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/layout-44c317d3520aab8e.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/page-33e307e1e914a553.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/(home)/layout-4421e85083373900.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/(home)/page-c75988fc008fc07a.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/chatroom/%5BroomId%5D/page-314d54e4d62d6237.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/chats/@tabs/before/page-cad9b93b3eadb665.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/chats/@tabs/closed/page-b86d978edbb0af28.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/chats/@tabs/confirm/page-b1f8b92086d19bdf.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/chats/@tabs/layout-6abd9c789f9fa50d.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/chats/@tabs/page-9e9351b46003c8d9.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/chats/@tabs/pending/page-ab10657bc54e25f8.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/chats/layout-ba06548ddfe3cc6a.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/mypage/cash/account/page-3a428fc2976da891.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/mypage/cash/exchange/page-292fe92c0c2c36c6.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/mypage/cash/history/page-4ad6d86aa2cffcdc.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/mypage/cash/page-7293965776456af8.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/mypage/menu/page-9b4ce6f69ec12300.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/mypage/page-aa9337b49d3fb760.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/mypage/preview/@tabs/information/page-246dfd1996fd0744.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/mypage/preview/@tabs/layout-c658d377c3177508.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/mypage/preview/@tabs/page-7469fed6a7d61b71.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/mypage/preview/@tabs/review/page-81107cc6ecf231be.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/mypage/preview/layout-0afe67987a11fd5f.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/mypage/profile/page-da7535384828687f.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/mypage/styles/%5BpostId%5D/page-02bc8c18ff85df3d.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/mypage/styles/edit/page-0a6db8570bae4764.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/partner/reqcoordi/%5BrequestId%5D/page-02740e334bed9617.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/(home)/layout-d3b70555c2c266d6.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/(home)/page-03f7faa0744b1ead.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/chatroom/%5BroomId%5D/page-a15af4b03ead4eeb.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/chats/@tabs/before/page-b40c641c6af369cc.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/chats/@tabs/closed/page-7ecd5f49bc8c71c3.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/chats/@tabs/confirm/page-3c2b498355ee6f43.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/chats/@tabs/layout-0f62b4634e2bceeb.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/chats/@tabs/page-1d2936915ac5a2e6.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/chats/@tabs/pending/page-91cc698cf75748fa.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/chats/layout-1f1009ef11dedf7e.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/confirm/page-2dcfedef03afbb67.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/profile/@tabs/information/page-a6bb47a53c84bfef.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/profile/@tabs/layout-b5fdd46447fdd0c2.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/profile/@tabs/page-63f567158c9c3aa2.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/profile/@tabs/review/page-e6dfc62f0962996e.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/profile/layout-be36d66beba358ab.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/reqcoordi/layout-7d2481a8ee347b22.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/reqcoordi/page-8c6730620c6edeee.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/coordinator/page-881e9020c703b517.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/mypage/page-732bbc5a3f7f85a8.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/mypage/profile/editbodyinfo/page-e0389b8cc97026f3.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/mypage/profile/editsize/page-5e7312f23607ae16.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/mypage/profile/editstyle/page-95abdc7b35e8c97b.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/mypage/profile/layout-b304769214e255ea.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/mypage/profile/page-751c3a6da2477eae.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/mypage/reqlist/layout-e7d5a159995f5841.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/mypage/reqlist/newreq/page-ab0d922283453a3e.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/mypage/reqlist/page-5d521a9e6027a646.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/payment/cash/charge/page-fa42ec8f38542f5a.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/payment/confirm/page-889b106b2b748c97.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/payment/page-71ee4f4fa52bbb7b.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/app/user/payment/result/page-84b3c58b7a49775a.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/fd9d1056-1d75990e62d644bb.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/framework-8e0e0f4a6b83a956.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/main-9f5f406f9515cf07.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/main-app-2cc063d09c3076b8.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/pages/_app-f870474a17b7f2fd.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/pages/_error-c66a4e8afc46f17b.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/chunks/polyfills-78c92fac7aa8fdd8.js",revision:"79330112775102f91e1010318bae2bd3"},{url:"/_next/static/chunks/webpack-ac4a612fe44dc3d3.js",revision:"oqa_-PHa8NyVJs8kFBQU9"},{url:"/_next/static/css/7bf773dac7d32baf.css",revision:"7bf773dac7d32baf"},{url:"/_next/static/css/d7e2247a4d8a8c0e.css",revision:"d7e2247a4d8a8c0e"},{url:"/_next/static/media/ff840cfebfb63b0c-s.p.woff2",revision:"302ec55f5b4320354ec6b35a53dead87"},{url:"/_next/static/oqa_-PHa8NyVJs8kFBQU9/_buildManifest.js",revision:"3e2d62a10f4d6bf0b92e14aecf7836f4"},{url:"/_next/static/oqa_-PHa8NyVJs8kFBQU9/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/bank/032.svg",revision:"a717346ebf03863ca777c31e5a5cd47b"},{url:"/firebase-messaging-sw.js",revision:"e5062b1654de7390726fd795cd9a59d4"},{url:"/icon-192x192.png",revision:"0846a77b2550492e534f7a1a7012eee8"},{url:"/icon-256x256.png",revision:"bac2fc762e53b57f2ce55e628e5c5be9"},{url:"/icon-384x384.png",revision:"e9d31a292c162deb204ee96297648157"},{url:"/icon-512x512.png",revision:"85feb3a273be45d763ea33b13ef941c8"},{url:"/icons/add-circle.svg",revision:"cb32e1a7838bd25298b2db63441097d5"},{url:"/icons/add-icon.svg",revision:"6f9683700bb6cdcd2feddb69e3b74c90"},{url:"/icons/alert.svg",revision:"53b457ffe1c1916e435d1cb6b0a98736"},{url:"/icons/blog.svg",revision:"55481482768728825523f8ea0fdc931e"},{url:"/icons/chats.svg",revision:"4105bffa4387a8c629b2679e1fcc9083"},{url:"/icons/delete-circle-red.svg",revision:"681fb4934a04d4071552589ee26ad0c3"},{url:"/icons/explore.svg",revision:"469da60dab0d375274bf8cf87f0bf51b"},{url:"/icons/facebook.svg",revision:"bf6860ea5647ee8e43edf4dce776777c"},{url:"/icons/hamburger.svg",revision:"cbfcd2b4b0845ebb32afe9f3c13de158"},{url:"/icons/hanger.svg",revision:"5130b0e1763a1e0516389a58be85ba18"},{url:"/icons/home.svg",revision:"c89088f6434a65d452e060c083099159"},{url:"/icons/instagram.svg",revision:"8c1d1ca2d3c2f7715d2adbc829f2869a"},{url:"/icons/kakaopay.svg",revision:"ff7e7cb122f01358878e98e294d9d134"},{url:"/icons/link.svg",revision:"2ed8638625cde7ffcbcfc2e456e547c8"},{url:"/icons/list-arrow.svg",revision:"bf0c12c878ec1ee4c55911008ba4e2ed"},{url:"/icons/menu-line.svg",revision:"3b4b04bdc82133b9142f5c883144f7af"},{url:"/icons/money-send.svg",revision:"f66333b18bdfbcd547aa051d74414099"},{url:"/icons/more-vertical.svg",revision:"efc0580ee3265c28a768f1dcb763a57d"},{url:"/icons/mypage.svg",revision:"7f3fd0a5b4015fee5c6257a7c068f4aa"},{url:"/icons/notifications.svg",revision:"f5557e4b6c5a9f39f0b74d0715d00c45"},{url:"/icons/npay.svg",revision:"a9f434e046ff13db5bc42461928854fd"},{url:"/icons/pig.svg",revision:"551c7bc96aeb36a25ff438f0597dd710"},{url:"/icons/ranking.svg",revision:"dde98b9bf0a72b0d1dbad0783983384e"},{url:"/icons/reminder.svg",revision:"452c22ae5b4e99ffe92dc5bf61a94196"},{url:"/icons/shop.svg",revision:"201a70812f3aabe13ab5a18e5fbea1c1"},{url:"/icons/smile-coin.svg",revision:"2521054e2e3b2fbf73097cb30eb25108"},{url:"/icons/tiktok.svg",revision:"174968eace91fd753881942b5e6279ca"},{url:"/icons/tosspay.svg",revision:"e2d26fb2d372b8e779dddb426d5f525c"},{url:"/icons/youtube.svg",revision:"15bfed8f7002d6ca54d323858ec05f8c"},{url:"/images/chat-add-file.svg",revision:"aff975e993ad7c4fc3ebb63cb8774e43"},{url:"/images/chat-send-button.svg",revision:"a2c1cb5c63655da881358e5b35070630"},{url:"/images/home.svg",revision:"b36b13ea11bc0c657091ebf3e4995fe4"},{url:"/images/mfc-logo.png",revision:"1509ef459129bc4ce20097d00fcdba46"},{url:"/images/my_image.jpg",revision:"2cc6b660848720a1e3f0cba59cc63bbb"},{url:"/images/partner-post.jpeg",revision:"600a5b6bbd785e3865ce3109e51e8cb0"},{url:"/images/partner-signup.svg",revision:"c8686aa20857250ff3c077962437b605"},{url:"/images/pencil-icon.svg",revision:"0077cf4128c546f06f5646eb24a03847"},{url:"/images/pencil.svg",revision:"7ec95ef405e963f8debfbf6fb423cea6"},{url:"/images/profile.png",revision:"b2186ebae126a379baa02be675111dc4"},{url:"/images/reference_image.jpg",revision:"3d3970ae7539268f3695be6595dde26c"},{url:"/images/test-profile.jpg",revision:"faeecadf6a5cec73918943eabf140f00"},{url:"/images/test.svg",revision:"9ee51ab61065d6c47ce4a8803a3976b4"},{url:"/images/user-signup-man.svg",revision:"0725c84c4c32396f4ccdbf96359b82d4"},{url:"/images/user-signup-woman.svg",revision:"ff804f663fe148de973c951c42a6147d"},{url:"/initial_page.svg",revision:"df58b339bac3d2e4699a930bc4aeb761"},{url:"/manifest.json",revision:"49377dff528b9ba7c1a926a3f5242dd3"},{url:"/next.svg",revision:"8e061864f388b47f33a1c3780831193e"},{url:"/vercel.svg",revision:"61c6b19abff40ea7acd577be818f3976"}],{ignoreURLParametersMatching:[/^utm_/,/^fbclid$/]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({response:e})=>e&&"opaqueredirect"===e.type?new Response(e.body,{status:200,statusText:"OK",headers:e.headers}):e}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:2592e3})]}),"GET"),e.registerRoute(/\/_next\/static.+\.js$/i,new e.CacheFirst({cacheName:"next-static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4|webm)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:48,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({sameOrigin:e,url:{pathname:a}})=>!(!e||a.startsWith("/api/auth/callback")||!a.startsWith("/api/"))),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({request:e,url:{pathname:a},sameOrigin:s})=>"1"===e.headers.get("RSC")&&"1"===e.headers.get("Next-Router-Prefetch")&&s&&!a.startsWith("/api/")),new e.NetworkFirst({cacheName:"pages-rsc-prefetch",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({request:e,url:{pathname:a},sameOrigin:s})=>"1"===e.headers.get("RSC")&&s&&!a.startsWith("/api/")),new e.NetworkFirst({cacheName:"pages-rsc",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:{pathname:e},sameOrigin:a})=>a&&!e.startsWith("/api/")),new e.NetworkFirst({cacheName:"pages",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({sameOrigin:e})=>!e),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET"),self.__WB_DISABLE_DEV_LOGS=!0}));
+if(!self.define){let e,s={};const a=(a,i)=>(a=new URL(a+".js",i).href,s[a]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=a,e.onload=s,document.head.appendChild(e)}else e=a,importScripts(a),s()})).then((()=>{let e=s[a];if(!e)throw new Error(`Module ${a} didn’t register its module`);return e})));self.define=(i,n)=>{const c=e||("document"in self?document.currentScript.src:"")||location.href;if(s[c])return;let r={};const t=e=>a(e,c),o={module:{uri:c},exports:r,require:t};s[c]=Promise.all(i.map((e=>o[e]||t(e)))).then((e=>(n(...e),r)))}}define(["./workbox-c06b064f"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/static/DUAg9XmHqh5Y9W5wVTJqH/_buildManifest.js",revision:"3e2d62a10f4d6bf0b92e14aecf7836f4"},{url:"/_next/static/DUAg9XmHqh5Y9W5wVTJqH/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/231-8cbc9f07fd7974e2.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/2433-a3c2272ce6e32cc3.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/2778-6a1712d23e611409.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/2967-86aef47a05f0281b.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/3771-c2dc1aac67d0345a.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/5190-88107600b5c8b2fc.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/6079-79ce56e79ed4ddb4.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/6111-fc803a4b575d4b6c.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/6612-1dc1b3cadee020c9.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/7116-60045fc3c0b0381f.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/7271-938511c36ff30d26.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/8173-275b4573b7cac1f5.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/8318a8be-eb73eb9ebc18b091.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/8866-0a1ff8780df5aa87.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/9646-15f0506087ab33c8.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/998-3aa0451c39cf29b9.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/(common)/(auth)/signin/page-211b7bed5ceedb6e.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/(common)/(auth)/signup/page-b3a0a794cf008435.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/(common)/explore/page-0533f035447ca065.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/(common)/ranking/page-c6703a6b471ff11d.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/(common)/selectrole/page-b78e3f88cc670408.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/_not-found/page-c51cc89ad3b69e1f.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/layout-44c317d3520aab8e.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/page-33e307e1e914a553.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/(home)/layout-4421e85083373900.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/(home)/page-c75988fc008fc07a.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/chatroom/%5BroomId%5D/page-314d54e4d62d6237.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/chats/@tabs/before/page-5f778c59b8ddecfa.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/chats/@tabs/closed/page-9b1d6aa9f120005c.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/chats/@tabs/confirm/page-0b4bc84793d7f7dc.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/chats/@tabs/layout-6abd9c789f9fa50d.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/chats/@tabs/page-9e9351b46003c8d9.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/chats/@tabs/pending/page-240e03caf0e05d7c.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/chats/layout-7d01b7bdf847181b.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/mypage/cash/account/page-3a428fc2976da891.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/mypage/cash/exchange/page-292fe92c0c2c36c6.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/mypage/cash/history/page-4ad6d86aa2cffcdc.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/mypage/cash/page-7293965776456af8.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/mypage/menu/page-9b4ce6f69ec12300.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/mypage/page-aa9337b49d3fb760.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/mypage/preview/@tabs/information/page-246dfd1996fd0744.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/mypage/preview/@tabs/layout-c658d377c3177508.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/mypage/preview/@tabs/page-3cb447923710b18d.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/mypage/preview/@tabs/review/page-f3db921b113b26ca.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/mypage/preview/layout-6d7ae67971388230.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/mypage/profile/page-da7535384828687f.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/mypage/styles/%5BpostId%5D/page-02bc8c18ff85df3d.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/mypage/styles/edit/page-0a6db8570bae4764.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/partner/reqcoordi/%5BrequestId%5D/page-02740e334bed9617.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/(home)/layout-e2a9748cfc0c1f14.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/(home)/page-03f7faa0744b1ead.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/chatroom/%5BroomId%5D/page-a15af4b03ead4eeb.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/chats/@tabs/before/page-7a4a02212bb92ace.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/chats/@tabs/closed/page-eee800d9aba65611.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/chats/@tabs/confirm/page-ced5aaa3a9a49751.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/chats/@tabs/layout-0f62b4634e2bceeb.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/chats/@tabs/page-1d2936915ac5a2e6.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/chats/@tabs/pending/page-b634e37758407822.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/chats/layout-3b557cbffaf06d30.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/confirm/page-2dcfedef03afbb67.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/profile/@tabs/information/page-a6bb47a53c84bfef.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/profile/@tabs/layout-b5fdd46447fdd0c2.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/profile/@tabs/page-63f567158c9c3aa2.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/profile/@tabs/review/page-2beff9ed341af809.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/profile/layout-e7ed819a9574b611.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/reqcoordi/layout-7d2481a8ee347b22.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/reqcoordi/page-8c6730620c6edeee.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/coordinator/page-f793d8bc0ed99037.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/mypage/page-732bbc5a3f7f85a8.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/mypage/profile/editbodyinfo/page-e0389b8cc97026f3.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/mypage/profile/editsize/page-5e7312f23607ae16.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/mypage/profile/editstyle/page-95abdc7b35e8c97b.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/mypage/profile/layout-b304769214e255ea.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/mypage/profile/page-751c3a6da2477eae.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/mypage/reqlist/layout-e7d5a159995f5841.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/mypage/reqlist/newreq/page-ab0d922283453a3e.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/mypage/reqlist/page-5d521a9e6027a646.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/payment/cash/charge/page-fa42ec8f38542f5a.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/app/user/payment/confirm/page-a0a4807f5852584d.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/fd9d1056-1d75990e62d644bb.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/framework-8e0e0f4a6b83a956.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/main-2d313bd0e4707c0b.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/main-app-2cc063d09c3076b8.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/pages/_app-f870474a17b7f2fd.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/pages/_error-c66a4e8afc46f17b.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/chunks/polyfills-78c92fac7aa8fdd8.js",revision:"79330112775102f91e1010318bae2bd3"},{url:"/_next/static/chunks/webpack-ac4a612fe44dc3d3.js",revision:"DUAg9XmHqh5Y9W5wVTJqH"},{url:"/_next/static/css/7bf773dac7d32baf.css",revision:"7bf773dac7d32baf"},{url:"/_next/static/css/d7e2247a4d8a8c0e.css",revision:"d7e2247a4d8a8c0e"},{url:"/_next/static/media/ff840cfebfb63b0c-s.p.woff2",revision:"302ec55f5b4320354ec6b35a53dead87"},{url:"/bank/032.svg",revision:"a717346ebf03863ca777c31e5a5cd47b"},{url:"/firebase-messaging-sw.js",revision:"e5062b1654de7390726fd795cd9a59d4"},{url:"/icon-192x192.png",revision:"0846a77b2550492e534f7a1a7012eee8"},{url:"/icon-256x256.png",revision:"bac2fc762e53b57f2ce55e628e5c5be9"},{url:"/icon-384x384.png",revision:"e9d31a292c162deb204ee96297648157"},{url:"/icon-512x512.png",revision:"85feb3a273be45d763ea33b13ef941c8"},{url:"/icons/add-circle.svg",revision:"cb32e1a7838bd25298b2db63441097d5"},{url:"/icons/add-icon.svg",revision:"6f9683700bb6cdcd2feddb69e3b74c90"},{url:"/icons/alert.svg",revision:"53b457ffe1c1916e435d1cb6b0a98736"},{url:"/icons/blog.svg",revision:"55481482768728825523f8ea0fdc931e"},{url:"/icons/chats.svg",revision:"4105bffa4387a8c629b2679e1fcc9083"},{url:"/icons/delete-circle-red.svg",revision:"681fb4934a04d4071552589ee26ad0c3"},{url:"/icons/explore.svg",revision:"469da60dab0d375274bf8cf87f0bf51b"},{url:"/icons/facebook.svg",revision:"bf6860ea5647ee8e43edf4dce776777c"},{url:"/icons/hamburger.svg",revision:"cbfcd2b4b0845ebb32afe9f3c13de158"},{url:"/icons/hanger.svg",revision:"5130b0e1763a1e0516389a58be85ba18"},{url:"/icons/home.svg",revision:"c89088f6434a65d452e060c083099159"},{url:"/icons/instagram.svg",revision:"8c1d1ca2d3c2f7715d2adbc829f2869a"},{url:"/icons/kakaopay.svg",revision:"ff7e7cb122f01358878e98e294d9d134"},{url:"/icons/link.svg",revision:"2ed8638625cde7ffcbcfc2e456e547c8"},{url:"/icons/list-arrow.svg",revision:"bf0c12c878ec1ee4c55911008ba4e2ed"},{url:"/icons/menu-line.svg",revision:"3b4b04bdc82133b9142f5c883144f7af"},{url:"/icons/money-send.svg",revision:"f66333b18bdfbcd547aa051d74414099"},{url:"/icons/more-vertical.svg",revision:"efc0580ee3265c28a768f1dcb763a57d"},{url:"/icons/mypage.svg",revision:"7f3fd0a5b4015fee5c6257a7c068f4aa"},{url:"/icons/notifications.svg",revision:"f5557e4b6c5a9f39f0b74d0715d00c45"},{url:"/icons/npay.svg",revision:"a9f434e046ff13db5bc42461928854fd"},{url:"/icons/pig.svg",revision:"551c7bc96aeb36a25ff438f0597dd710"},{url:"/icons/ranking.svg",revision:"dde98b9bf0a72b0d1dbad0783983384e"},{url:"/icons/reminder.svg",revision:"452c22ae5b4e99ffe92dc5bf61a94196"},{url:"/icons/shop.svg",revision:"201a70812f3aabe13ab5a18e5fbea1c1"},{url:"/icons/smile-coin.svg",revision:"2521054e2e3b2fbf73097cb30eb25108"},{url:"/icons/tiktok.svg",revision:"174968eace91fd753881942b5e6279ca"},{url:"/icons/tosspay.svg",revision:"e2d26fb2d372b8e779dddb426d5f525c"},{url:"/icons/youtube.svg",revision:"15bfed8f7002d6ca54d323858ec05f8c"},{url:"/images/chat-add-file.svg",revision:"aff975e993ad7c4fc3ebb63cb8774e43"},{url:"/images/chat-send-button.svg",revision:"a2c1cb5c63655da881358e5b35070630"},{url:"/images/home.svg",revision:"b36b13ea11bc0c657091ebf3e4995fe4"},{url:"/images/mfc-logo.png",revision:"1509ef459129bc4ce20097d00fcdba46"},{url:"/images/my_image.jpg",revision:"2cc6b660848720a1e3f0cba59cc63bbb"},{url:"/images/partner-post.jpeg",revision:"600a5b6bbd785e3865ce3109e51e8cb0"},{url:"/images/partner-signup.svg",revision:"c8686aa20857250ff3c077962437b605"},{url:"/images/pencil-icon.svg",revision:"0077cf4128c546f06f5646eb24a03847"},{url:"/images/pencil.svg",revision:"7ec95ef405e963f8debfbf6fb423cea6"},{url:"/images/profile.png",revision:"b2186ebae126a379baa02be675111dc4"},{url:"/images/reference_image.jpg",revision:"3d3970ae7539268f3695be6595dde26c"},{url:"/images/test-profile.jpg",revision:"faeecadf6a5cec73918943eabf140f00"},{url:"/images/test.svg",revision:"9ee51ab61065d6c47ce4a8803a3976b4"},{url:"/images/user-signup-man.svg",revision:"0725c84c4c32396f4ccdbf96359b82d4"},{url:"/images/user-signup-woman.svg",revision:"ff804f663fe148de973c951c42a6147d"},{url:"/initial_page.svg",revision:"df58b339bac3d2e4699a930bc4aeb761"},{url:"/manifest.json",revision:"49377dff528b9ba7c1a926a3f5242dd3"},{url:"/next.svg",revision:"8e061864f388b47f33a1c3780831193e"},{url:"/vercel.svg",revision:"61c6b19abff40ea7acd577be818f3976"}],{ignoreURLParametersMatching:[/^utm_/,/^fbclid$/]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({response:e})=>e&&"opaqueredirect"===e.type?new Response(e.body,{status:200,statusText:"OK",headers:e.headers}):e}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:2592e3})]}),"GET"),e.registerRoute(/\/_next\/static.+\.js$/i,new e.CacheFirst({cacheName:"next-static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4|webm)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:48,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({sameOrigin:e,url:{pathname:s}})=>!(!e||s.startsWith("/api/auth/callback")||!s.startsWith("/api/"))),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({request:e,url:{pathname:s},sameOrigin:a})=>"1"===e.headers.get("RSC")&&"1"===e.headers.get("Next-Router-Prefetch")&&a&&!s.startsWith("/api/")),new e.NetworkFirst({cacheName:"pages-rsc-prefetch",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({request:e,url:{pathname:s},sameOrigin:a})=>"1"===e.headers.get("RSC")&&a&&!s.startsWith("/api/")),new e.NetworkFirst({cacheName:"pages-rsc",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:{pathname:e},sameOrigin:s})=>s&&!e.startsWith("/api/")),new e.NetworkFirst({cacheName:"pages",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({sameOrigin:e})=>!e),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET"),self.__WB_DISABLE_DEV_LOGS=!0}));
